Remove synchronous API.
Remove NotificationHandler support in favour of stream.
