+++ src/api/mod.rs
@@ -12,22 +12,23 @@
 // Copyright (c) 2014 The Rust Project Developers
 
 mod adapter_manager;
-#[cfg(feature = "async")]
-pub mod async_api;
 mod bdaddr;
 pub mod bleuuid;
 
-use crate::Result;
 pub use adapter_manager::AdapterManager;
+
+use crate::Result;
+use async_trait::async_trait;
 use bitflags::bitflags;
+use futures::stream::Stream;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 #[cfg(feature = "serde")]
 use serde_cr as serde;
-use std::sync::mpsc::Receiver;
 use std::{
     collections::{BTreeSet, HashMap},
     fmt::{self, Debug, Display, Formatter},
+    pin::Pin,
 };
 use uuid::Uuid;
 
@@ -86,10 +87,6 @@ pub struct ValueNotification {
     /// The new value of the handle.
     pub value: Vec<u8>,
 }
-
-pub type Callback<T> = Box<dyn Fn(Result<T>) + Send>;
-
-pub type NotificationHandler = Box<dyn FnMut(ValueNotification) + Send>;
 
 bitflags! {
     /// A set of properties that indicate what operations are supported by a Characteristic.
@@ -187,63 +184,56 @@ pub enum WriteType {
 /// Peripheral is the device that you would like to communicate with (the "server" of BLE). This
 /// struct contains both the current state of the device (its properties, characteristics, etc.)
 /// as well as functions for communication.
+#[async_trait]
 pub trait Peripheral: Send + Sync + Clone + Debug {
-    /// Returns the address of the peripheral.
+    /// Returns the MAC address of the peripheral.
     fn address(&self) -> BDAddr;
 
     /// Returns the set of properties associated with the peripheral. These may be updated over time
     /// as additional advertising reports are received.
-    fn properties(&self) -> PeripheralProperties;
+    async fn properties(&self) -> Result<PeripheralProperties>;
 
     /// The set of characteristics we've discovered for this device. This will be empty until
     /// `discover_characteristics` is called.
     fn characteristics(&self) -> BTreeSet<Characteristic>;
 
     /// Returns true iff we are currently connected to the device.
-    fn is_connected(&self) -> bool;
+    async fn is_connected(&self) -> Result<bool>;
 
-    /// Creates a connection to the device. This is a synchronous operation; if this method returns
-    /// Ok there has been successful connection. Note that peripherals allow only one connection at
-    /// a time. Operations that attempt to communicate with a device will fail until it is connected.
-    fn connect(&self) -> Result<()>;
+    /// Creates a connection to the device. If this method returns Ok there has been successful
+    /// connection. Note that peripherals allow only one connection at a time. Operations that
+    /// attempt to communicate with a device will fail until it is connected.
+    async fn connect(&self) -> Result<()>;
 
-    /// Terminates a connection to the device. This is a synchronous operation.
-    fn disconnect(&self) -> Result<()>;
+    /// Terminates a connection to the device.
+    async fn disconnect(&self) -> Result<()>;
 
-    /// Discovers all characteristics for the device. This is a synchronous operation.
-    fn discover_characteristics(&self) -> Result<Vec<Characteristic>>;
+    /// Discovers all characteristics for the device.
+    async fn discover_characteristics(&self) -> Result<Vec<Characteristic>>;
 
-    /// Write some data to the characteristic. Returns an error if the write couldn't be send or (in
+    /// Write some data to the characteristic. Returns an error if the write couldn't be sent or (in
     /// the case of a write-with-response) if the device returns an error.
-    fn write(
+    async fn write(
         &self,
         characteristic: &Characteristic,
         data: &[u8],
         write_type: WriteType,
     ) -> Result<()>;
 
-    /// Sends a request (read) to the device. Synchronously returns either an error if the request
-    /// was not accepted or the response from the device.
-    fn read(&self, characteristic: &Characteristic) -> Result<Vec<u8>>;
-
-    /// Sends a read-by-type request to device for the range of handles covered by the
-    /// characteristic and for the specified declaration UUID. See
-    /// [here](https://www.bluetooth.com/specifications/gatt/declarations) for valid UUIDs.
-    /// Synchronously returns either an error or the device response.
-    fn read_by_type(&self, characteristic: &Characteristic, uuid: Uuid) -> Result<Vec<u8>>;
+    /// Sends a read request to the device. Returns either an error if the request was not accepted
+    /// or the response from the device.
+    async fn read(&self, characteristic: &Characteristic) -> Result<Vec<u8>>;
 
     /// Enables either notify or indicate (depending on support) for the specified characteristic.
-    /// This is a synchronous call.
-    fn subscribe(&self, characteristic: &Characteristic) -> Result<()>;
+    async fn subscribe(&self, characteristic: &Characteristic) -> Result<()>;
 
     /// Disables either notify or indicate (depending on support) for the specified characteristic.
-    /// This is a synchronous call.
-    fn unsubscribe(&self, characteristic: &Characteristic) -> Result<()>;
+    async fn unsubscribe(&self, characteristic: &Characteristic) -> Result<()>;
 
-    /// Registers a handler that will be called when value notification messages are received from
-    /// the device. This method should only be used after a connection has been established. Note
-    /// that the handler will be called in a common thread, so it should not block.
-    fn on_notification(&self, handler: NotificationHandler);
+    /// Returns a stream of notifications for characteristic value updates. The stream will receive
+    /// a notification when a value notification or indication is received from the device. This
+    /// method should only be used after a connection has been established.
+    async fn notifications(&self) -> Result<Pin<Box<dyn Stream<Item = ValueNotification>>>>;
 }
 
 #[cfg_attr(
@@ -276,40 +266,38 @@ pub enum CentralEvent {
 }
 
 /// Central is the "client" of BLE. It's able to scan for and establish connections to peripherals.
+#[async_trait]
 pub trait Central: Send + Sync + Clone {
     type Peripheral: Peripheral;
 
-    /// Retreive the Event [Receiver] for the event channel. This channel
-    /// receiver will receive notifications when events occur for this Central
-    /// module. As this uses an std::channel which cannot be cloned, after the
-    /// first call (which will contain Some<Receiver<CentralEvent>>), all
-    /// subsequent calls will return None. See [`Event`](enum.CentralEvent.html)
-    /// for the full set of events returned.
-    fn event_receiver(&self) -> Option<Receiver<CentralEvent>>;
+    /// Retreive a stream of `CentralEvent`s. This stream will receive notifications when events
+    /// occur for this Central module. See [`CentralEvent`](enum.CentralEvent.html) for the full set
+    /// of possible events.
+    async fn events(&self) -> Result<Pin<Box<dyn Stream<Item = CentralEvent>>>>;
 
     /// Starts a scan for BLE devices. This scan will generally continue until explicitly stopped,
-    /// although this may depend on your bluetooth adapter. Discovered devices will be announced
-    /// to subscribers of `on_event` and will be available via `peripherals()`.
-    fn start_scan(&self) -> Result<()>;
+    /// although this may depend on your Bluetooth adapter. Discovered devices will be announced
+    /// to subscribers of `events` and will be available via `peripherals()`.
+    async fn start_scan(&self) -> Result<()>;
 
     /// Control whether to use active or passive scan mode to find BLE devices. Active mode scan
-    /// notifies advertises about the scan, whereas passive scan only receives data from the
+    /// notifies advertisers about the scan, whereas passive scan only receives data from the
     /// advertiser. Defaults to use active mode.
-    fn active(&self, enabled: bool);
+    async fn active(&self, enabled: bool);
 
-    /// Control whether to filter multiple advertisements by the same peer device. Receving
-    /// can be useful for some applications. E.g. when using scan to collect information from
-    /// beacons that update data frequently. Defaults to filter duplicate advertisements.
-    fn filter_duplicates(&self, enabled: bool);
+    /// Control whether to filter multiple advertisements by the same peer device. Duplicates can be
+    /// useful for some applications, e.g. when using a scan to collect information from beacons
+    /// that update data frequently. Defaults to filter duplicate advertisements.
+    async fn filter_duplicates(&self, enabled: bool);
 
     /// Stops scanning for BLE devices.
-    fn stop_scan(&self) -> Result<()>;
+    async fn stop_scan(&self) -> Result<()>;
 
     /// Returns the list of [`Peripherals`](trait.Peripheral.html) that have been discovered so far.
     /// Note that this list may contain peripherals that are no longer available.
-    fn peripherals(&self) -> Vec<Self::Peripheral>;
+    async fn peripherals(&self) -> Result<Vec<Self::Peripheral>>;
 
     /// Returns a particular [`Peripheral`](trait.Peripheral.html) by its address if it has been
     /// discovered.
-    fn peripheral(&self, address: BDAddr) -> Option<Self::Peripheral>;
+    async fn peripheral(&self, address: BDAddr) -> Result<Self::Peripheral>;
 }
