--- src/corebluetooth/async_api/mod.rs
@@ -1,3 +1,0 @@
-pub mod adapter;
-pub mod manager;
-pub mod peripheral;
