+++ src/corebluetooth/peripheral.rs
@@ -13,33 +13,34 @@ use super::{
 };
 use crate::{
     api::{
-        AdapterManager, AddressType, BDAddr, CentralEvent, Characteristic, NotificationHandler,
-        Peripheral as ApiPeripheral, PeripheralProperties, ValueNotification, WriteType,
+        self, AdapterManager, AddressType, BDAddr, CentralEvent, Characteristic,
+        PeripheralProperties, ValueNotification, WriteType,
     },
     common::util,
     Error, Result,
 };
 use async_std::task;
-use futures::channel::mpsc::{Receiver, SendError, Sender};
+use async_trait::async_trait;
+use futures::channel::mpsc::{self, Receiver, SendError, Sender, UnboundedSender};
 use futures::sink::SinkExt;
-use futures::stream::StreamExt;
+use futures::stream::{Stream, StreamExt};
 use log::{debug, error, info};
 use std::{
     collections::{BTreeSet, HashMap},
     fmt::{self, Debug, Display, Formatter},
-    iter::FromIterator,
+    pin::Pin,
     sync::{Arc, Mutex},
 };
 use uuid::Uuid;
 
 #[derive(Clone)]
 pub struct Peripheral {
-    notification_handlers: Arc<Mutex<Vec<NotificationHandler>>>,
+    notification_senders: Arc<Mutex<Vec<UnboundedSender<ValueNotification>>>>,
     manager: AdapterManager<Self>,
-    pub(crate) uuid: Uuid,
-    pub(crate) characteristics: Arc<Mutex<BTreeSet<Characteristic>>>,
-    pub(crate) properties: Arc<Mutex<PeripheralProperties>>,
-    pub(crate) message_sender: Sender<CoreBluetoothMessage>,
+    uuid: Uuid,
+    characteristics: Arc<Mutex<BTreeSet<Characteristic>>>,
+    properties: Arc<Mutex<PeripheralProperties>>,
+    message_sender: Sender<CoreBluetoothMessage>,
     // We're not actually holding a peripheral object here, that's held out in
     // the objc thread. We'll just communicate with it through our
     // receiver/sender pair.
@@ -68,8 +69,8 @@ impl Peripheral {
             discovery_count: 1,
             has_scan_response: true,
         }));
-        let notification_handlers = Arc::new(Mutex::new(Vec::<NotificationHandler>::new()));
-        let nh_clone = notification_handlers.clone();
+        let notification_senders = Arc::new(Mutex::new(Vec::new()));
+        let ns_clone = notification_senders.clone();
         let p_clone = properties.clone();
         let m_clone = manager.clone();
         task::spawn(async move {
@@ -77,8 +78,8 @@ impl Peripheral {
             loop {
                 match event_receiver.next().await {
                     Some(CBPeripheralEvent::Notification(uuid, data)) => {
-                        util::invoke_handlers(
-                            &nh_clone,
+                        util::send_notification(
+                            &ns_clone,
                             &ValueNotification {
                                 uuid,
                                 handle: None,
@@ -126,7 +127,7 @@ impl Peripheral {
             properties,
             manager,
             characteristics: Arc::new(Mutex::new(BTreeSet::new())),
-            notification_handlers,
+            notification_senders,
             uuid,
             message_sender,
         }
@@ -159,176 +160,138 @@ impl Debug for Peripheral {
     }
 }
 
-impl ApiPeripheral for Peripheral {
-    /// Returns the address of the peripheral.
+#[async_trait]
+impl api::Peripheral for Peripheral {
     fn address(&self) -> BDAddr {
         self.properties.lock().unwrap().address
     }
 
-    /// Returns the set of properties associated with the peripheral. These may be updated over time
-    /// as additional advertising reports are received.
-    fn properties(&self) -> PeripheralProperties {
-        self.properties.lock().unwrap().clone()
+    async fn properties(&self) -> Result<PeripheralProperties> {
+        Ok(self.properties.lock().unwrap().clone())
     }
 
-    /// The set of characteristics we've discovered for this device. This will be empty until
-    /// `discover_characteristics` is called.
     fn characteristics(&self) -> BTreeSet<Characteristic> {
         self.characteristics.lock().unwrap().clone()
     }
 
-    /// Returns true iff we are currently connected to the device.
-    fn is_connected(&self) -> bool {
-        false
+    async fn is_connected(&self) -> Result<bool> {
+        // TODO
+        Ok(false)
     }
 
-    /// Creates a connection to the device. This is a synchronous operation; if this method returns
-    /// Ok there has been successful connection. Note that peripherals allow only one connection at
-    /// a time. Operations that attempt to communicate with a device will fail until it is connected.
-    fn connect(&self) -> Result<()> {
-        info!("Trying device connect!");
-        task::block_on(async {
-            let mut message_sender = self.message_sender.clone();
-            let fut = CoreBluetoothReplyFuture::default();
-            message_sender
-                .send(CoreBluetoothMessage::ConnectDevice(
-                    self.uuid,
-                    fut.get_state_clone(),
-                ))
-                .await?;
-            match fut.await {
-                CoreBluetoothReply::Connected(chars) => {
-                    *(self.characteristics.lock().unwrap()) = chars;
-                    self.emit(CentralEvent::DeviceConnected(
-                        self.properties.lock().unwrap().address,
-                    ));
-                }
-                _ => panic!("Shouldn't get anything but connected!"),
+    async fn connect(&self) -> Result<()> {
+        let fut = CoreBluetoothReplyFuture::default();
+        self.message_sender
+            .to_owned()
+            .send(CoreBluetoothMessage::ConnectDevice(
+                self.uuid,
+                fut.get_state_clone(),
+            ))
+            .await?;
+        match fut.await {
+            CoreBluetoothReply::Connected(chars) => {
+                *(self.characteristics.lock().unwrap()) = chars;
+                self.emit(CentralEvent::DeviceConnected(
+                    self.properties.lock().unwrap().address,
+                ));
             }
-            info!("Device connected!");
-            Ok(())
-        })
-    }
-
-    /// Terminates a connection to the device. This is a synchronous operation.
-    fn disconnect(&self) -> Result<()> {
+            _ => panic!("Shouldn't get anything but connected!"),
+        }
+        info!("Device connected!");
         Ok(())
     }
 
-    /// Discovers all characteristics for the device. This is a synchronous operation.
-    fn discover_characteristics(&self) -> Result<Vec<Characteristic>> {
-        let chrs = self.characteristics.lock().unwrap().clone();
-        let v = Vec::from_iter(chrs.into_iter());
-        Ok(v)
+    async fn disconnect(&self) -> Result<()> {
+        // TODO
+        Ok(())
     }
 
-    /// Write some data to the characteristic. Returns an error if the write couldn't be send or (in
-    /// the case of a write-with-response) if the device returns an error.
-    fn write(
+    async fn discover_characteristics(&self) -> Result<Vec<Characteristic>> {
+        let characteristics = self.characteristics.lock().unwrap().clone();
+        Ok(characteristics.into_iter().collect())
+    }
+
+    async fn write(
         &self,
         characteristic: &Characteristic,
         data: &[u8],
         write_type: WriteType,
     ) -> Result<()> {
-        task::block_on(async {
-            let mut message_sender = self.message_sender.clone();
-            let fut = CoreBluetoothReplyFuture::default();
-            message_sender
-                .send(CoreBluetoothMessage::WriteValue(
-                    self.uuid,
-                    characteristic.uuid,
-                    Vec::from(data),
-                    write_type,
-                    fut.get_state_clone(),
-                ))
-                .await?;
-            match fut.await {
-                CoreBluetoothReply::Ok => {}
-                reply => panic!("Unexpected reply: {:?}", reply),
+        let fut = CoreBluetoothReplyFuture::default();
+        self.message_sender
+            .to_owned()
+            .send(CoreBluetoothMessage::WriteValue(
+                self.uuid,
+                characteristic.uuid,
+                Vec::from(data),
+                write_type,
+                fut.get_state_clone(),
+            ))
+            .await?;
+        match fut.await {
+            CoreBluetoothReply::Ok => {}
+            reply => panic!("Unexpected reply: {:?}", reply),
+        }
+        Ok(())
+    }
+
+    async fn read(&self, characteristic: &Characteristic) -> Result<Vec<u8>> {
+        let fut = CoreBluetoothReplyFuture::default();
+        self.message_sender
+            .to_owned()
+            .send(CoreBluetoothMessage::ReadValue(
+                self.uuid,
+                characteristic.uuid,
+                fut.get_state_clone(),
+            ))
+            .await?;
+        match fut.await {
+            CoreBluetoothReply::ReadResult(chars) => Ok(chars),
+            _ => {
+                panic!("Shouldn't get anything but read result!");
             }
-            Ok(())
-        })
+        }
     }
 
-    /// Sends a read-by-type request to device for the range of handles covered by the
-    /// characteristic and for the specified declaration UUID. See
-    /// [here](https://www.bluetooth.com/specifications/gatt/declarations) for valid UUIDs.
-    /// Synchronously returns either an error or the device response.
-    fn read_by_type(&self, _characteristic: &Characteristic, _uuid: Uuid) -> Result<Vec<u8>> {
-        Err(Error::NotSupported("read_by_type".into()))
+    async fn subscribe(&self, characteristic: &Characteristic) -> Result<()> {
+        let fut = CoreBluetoothReplyFuture::default();
+        self.message_sender
+            .to_owned()
+            .send(CoreBluetoothMessage::Subscribe(
+                self.uuid,
+                characteristic.uuid,
+                fut.get_state_clone(),
+            ))
+            .await?;
+        match fut.await {
+            CoreBluetoothReply::Ok => info!("subscribed!"),
+            _ => panic!("Didn't subscribe!"),
+        }
+        Ok(())
     }
 
-    /// Enables either notify or indicate (depending on support) for the specified characteristic.
-    /// This is a synchronous call.
-    fn subscribe(&self, characteristic: &Characteristic) -> Result<()> {
-        info!("Trying to subscribe!");
-        task::block_on(async {
-            let mut message_sender = self.message_sender.clone();
-            let fut = CoreBluetoothReplyFuture::default();
-            message_sender
-                .send(CoreBluetoothMessage::Subscribe(
-                    self.uuid,
-                    characteristic.uuid,
-                    fut.get_state_clone(),
-                ))
-                .await?;
-            match fut.await {
-                CoreBluetoothReply::Ok => info!("subscribed!"),
-                _ => panic!("Didn't subscribe!"),
-            }
-            Ok(())
-        })
+    async fn unsubscribe(&self, characteristic: &Characteristic) -> Result<()> {
+        let fut = CoreBluetoothReplyFuture::default();
+        self.message_sender
+            .to_owned()
+            .send(CoreBluetoothMessage::Unsubscribe(
+                self.uuid,
+                characteristic.uuid,
+                fut.get_state_clone(),
+            ))
+            .await?;
+        match fut.await {
+            CoreBluetoothReply::Ok => {}
+            _ => panic!("Didn't unsubscribe!"),
+        }
+        Ok(())
     }
 
-    /// Disables either notify or indicate (depending on support) for the specified characteristic.
-    /// This is a synchronous call.
-    fn unsubscribe(&self, characteristic: &Characteristic) -> Result<()> {
-        info!("Trying to unsubscribe!");
-        task::block_on(async {
-            let mut message_sender = self.message_sender.clone();
-            let fut = CoreBluetoothReplyFuture::default();
-            message_sender
-                .send(CoreBluetoothMessage::Unsubscribe(
-                    self.uuid,
-                    characteristic.uuid,
-                    fut.get_state_clone(),
-                ))
-                .await?;
-            match fut.await {
-                CoreBluetoothReply::Ok => {}
-                _ => panic!("Didn't unsubscribe!"),
-            }
-            Ok(())
-        })
-    }
-
-    /// Registers a handler that will be called when value notification messages are received from
-    /// the device. This method should only be used after a connection has been established. Note
-    /// that the handler will be called in a common thread, so it should not block.
-    fn on_notification(&self, handler: NotificationHandler) {
-        let mut list = self.notification_handlers.lock().unwrap();
-        list.push(handler);
-    }
-
-    fn read(&self, characteristic: &Characteristic) -> Result<Vec<u8>> {
-        info!("Trying read!");
-        task::block_on(async {
-            let mut message_sender = self.message_sender.clone();
-            let fut = CoreBluetoothReplyFuture::default();
-            message_sender
-                .send(CoreBluetoothMessage::ReadValue(
-                    self.uuid,
-                    characteristic.uuid,
-                    fut.get_state_clone(),
-                ))
-                .await?;
-            match fut.await {
-                CoreBluetoothReply::ReadResult(chars) => Ok(chars),
-                _ => {
-                    panic!("Shouldn't get anything but read result!");
-                }
-            }
-        })
+    async fn notifications(&self) -> Result<Pin<Box<dyn Stream<Item = ValueNotification>>>> {
+        let (sender, receiver) = mpsc::unbounded();
+        let mut senders = self.notification_senders.lock().unwrap();
+        senders.push(sender);
+        Ok(Box::pin(receiver))
     }
 }
