+++ src/lib.rs
@@ -20,8 +20,7 @@
 //! An example of how to use the library to control some BLE smart lights:
 //!
 //! ```rust,no_run
-//! use btleplug::api::async_api::{Central, Peripheral as _};
-//! use btleplug::api::{bleuuid::uuid_from_u16, WriteType};
+//! use btleplug::api::{bleuuid::uuid_from_u16, Central, Peripheral as _, WriteType};
 //! use btleplug::platform::{Adapter, Manager, Peripheral};
 //! use rand::{Rng, thread_rng};
 //! use std::error::Error;
@@ -96,7 +95,7 @@ use std::result;
 use std::time::Duration;
 
 pub mod api;
-#[cfg(all(target_os = "linux", feature = "async"))]
+#[cfg(target_os = "linux")]
 mod bluez;
 mod common;
 #[cfg(any(target_os = "macos", target_os = "ios"))]
